Consolidate tests into a single directory
