+++ containers/ecr-viewer/src/app/tests/components/Demographics.test.tsx
@@ -1,6 +1,6 @@
 import { render } from "@testing-library/react";
 import { axe } from "jest-axe";
-import Demographics from "../Demographics";
+import Demographics from "../../view-data/components/Demographics";
 
 describe("Demographics", () => {
   let container: HTMLElement;
