+++ containers/ecr-viewer/src/app/tests/components/LabInfo.test.tsx
